Don't use HashMap::entry API
In >99% of cases, AggMaps will contain the entry we're trying to add a
value to. The Entry API requires owned data in case it needs to insert,
so we were doing pointless string allocations for it.

Even with stack-allocated strings it ended up causing a bunch of useless
memcpys. This much simpler solution just checks for the strings by
reference, and if we need to insert, then it allocates a new String to
put in the map.